--- api/configs/dev_environment.example
@@ -1,25 +1,0 @@
-#!/usr/bin/env bash
-
-# Schema:
-# ENVIRONMENT_KEY=Value
-#
-# Please copy `./api/configs/dev_environment.example` to `./api/configs/dev_environment`
-# and fill in the missing information!
-
-LOGLEVEL=DEBUG
-FLASK_DEBUG=1
-
-# origin to allow requests from: as regular expression or string.
-# Include both the schema and the port (if not 80).
-# Please provide the url to the frontend.
-# default: "*"
-CORS_URL_REGEX="http://localhost:3000"
-
-GEOSERVER_WORKSPACE=CUT
-GEOSERVER_ADMIN_USER=admin
-GEOSERVER_ADMIN_PASSWORD=geoserver
-GEOSERVER_BASE_URL=http://geoserver:8080/geoserver
-
-# The API_SERVER_URL is only used to return the complete URL in the result of the job details as specified in OGC.
-# Should be the base url to the api.
-API_SERVER_URL=localhost:3000
