--- api/configs/prod_environment.example
@@ -1,25 +1,0 @@
-#!/usr/bin/env bash
-
-# Schema:
-# ENVIRONMENT_KEY=Value
-#
-# Please copy `./.env.example` to `./.env`
-# and fill in the missing information!
-LOGLEVEL=WARN
-FLASK_DEBUG=0
-NUMBER_OF_WORKERS=1
-
-# origin to allow requests from: as regular expression or string.
-# Include both the schema and the port (if not 80), e.g. https://example.com:3000
-# Please provide the url to the frontend.
-# default: "*"
-CORS_URL_REGEX="http://localhost:3000"
-
-GEOSERVER_WORKSPACE=CUT
-GEOSERVER_ADMIN_USER=admin
-GEOSERVER_ADMIN_PASSWORD=geoserver
-GEOSERVER_BASE_URL=http://geoserver:8080/geoserver
-
-# The SERVER_URL is only used to return the complete URL in the result of the job details as specified in OGC.
-# Should be the base url to the api.
-API_SERVER_URL=localhost:3000
