--- geoserver/configs/geoserver.example
@@ -1,37 +1,0 @@
-#!/usr/bin/env bash
-
-# Schema:
-# ENVIRONMENT_KEY=Value
-#
-# Please copy `./geoserver.example` to `./geoserver`
-# and fill in the missing information!
-
-GEOSERVER_ADMIN_USER=
-GEOSERVER_ADMIN_PASSWORD=
-
-GEOSERVER_DATA_DIR=
-GEOWEBCACHE_CACHE_DIR=
-INITIAL_MEMORY=
-MAXIMUM_MEMORY=
-STABLE_EXTENSIONS=
-COMMUNITY_EXTENSIONS=
-
-POSTGRES_USER=
-POSTGRES_PASSWORD=
-POSTGRES_DB=
-
-# for development
-# GEOSERVER_ADMIN_USER=admin
-# GEOSERVER_ADMIN_PASSWORD=geoserver
-
-# GEOSERVER_DATA_DIR=/opt/geoserver/data_dir
-# GEOWEBCACHE_CACHE_DIR=/opt/geoserver/cache_dir
-# INITIAL_MEMORY=2g
-# MAXIMUM_MEMORY=8g
-# STABLE_EXTENSIONS=[]
-# COMMUNITY_EXTENSIONS=[]
-# EXISTING_DATA_DIR=true
-
-# POSTGRES_USER=cut_dev
-# POSTGRES_PASSWORD=postgres
-# POSTGRES_DB=cut_dev
