--- geoserver/configs/postgis.example
@@ -1,22 +1,0 @@
-#!/usr/bin/env bash
-
-# Schema:
-# ENVIRONMENT_KEY=Value
-#
-# Please copy `./postgis.example` to `./postgis`
-# and fill in the missing information!
-
-POSTGRES_DB=
-POSTGRES_USER=
-POSTGRES_PASSWORD=
-PGDATA=
-POSTGRES_HOST=
-POSTGRES_PORT=
-
-# for development i.e.
-# POSTGRES_DB=cut_dev
-# POSTGRES_USER=postgres
-# POSTGRES_PASSWORD=postgres
-# PGDATA=/var/lib/postgresql/data
-# POSTGRES_HOST=postgis
-# POSTGRES_PORT=5432
