+++ src/ump/api/db_handler.py
@@ -1,4 +1,4 @@
-import configs.config as config
+import ump.config as config
 import psycopg2 as db
 from psycopg2.extras import RealDictCursor
